+++ tests/BaseTest.php
@@ -40,6 +40,21 @@ abstract class BaseTest extends TestCase {
     protected $api;
 
     /**
+     * @var \bheisig\idoitapi\CMDBObject
+     */
+    protected $cmdbObject;
+
+    /**
+     * @var \bheisig\idoitapi\CMDBObjects
+     */
+    protected $cmdbObjects;
+
+    /**
+     * @var \bheisig\idoitapi\CMDBCategory
+     */
+    protected $cmdbCategory;
+
+    /**
      * Information about this project
      *
      * @var array
@@ -73,6 +88,10 @@ abstract class BaseTest extends TestCase {
             'password' => $GLOBALS['password']
         ]);
 
+        $this->cmdbObject = new CMDBObject($this->api);
+        $this->cmdbObjects = new CMDBObjects($this->api);
+        $this->cmdbCategory = new CMDBCategory($this->api);
+
         $composerFile = __DIR__ . '/../composer.json';
 
         if (is_readable($composerFile)) {
@@ -88,9 +107,7 @@ abstract class BaseTest extends TestCase {
      * @throws \Exception
      */
     protected function createServer(): int {
-        $cmdbObject = new CMDBObject($this->api);
-
-        return $cmdbObject->create(
+        return $this->cmdbObject->create(
             'C__OBJTYPE__SERVER',
             $this->generateRandomString()
         );
@@ -104,9 +121,6 @@ abstract class BaseTest extends TestCase {
      * @throws \Exception
      */
     protected function createPerson(): array {
-        $cmdbObject = new CMDBObject($this->api);
-        $cmdbCategory = new CMDBCategory($this->api);
-
         $firstName = substr($this->generateRandomString(), 0, 10);
         $lastName = substr($this->generateRandomString(), 0, 10);
         $email = sprintf(
@@ -115,12 +129,12 @@ abstract class BaseTest extends TestCase {
             $lastName
         );
 
-        $personID = $cmdbObject->create(
+        $personID = $this->cmdbObject->create(
             'C__OBJTYPE__PERSON',
             $firstName . ' ' . $lastName
         );
 
-        $cmdbCategory->create(
+        $this->cmdbCategory->create(
             $personID,
             'C__CATG__MAIL_ADDRESSES',
             [
@@ -146,9 +160,7 @@ abstract class BaseTest extends TestCase {
      * @throws \Exception
      */
     protected function createWorkstation(): int {
-        $cmdbObject = new CMDBObject($this->api);
-
-        $workstationID = $cmdbObject->create(
+        $workstationID = $this->cmdbObject->create(
             'C__OBJTYPE__WORKSTATION',
             $this->generateRandomString()
         );
@@ -172,9 +184,7 @@ abstract class BaseTest extends TestCase {
      * @throws \Exception
      */
     protected function addPersonToWorkstation(int $personID, int $workstationID): int {
-        $cmdbCategory = new CMDBCategory($this->api);
-
-        return $cmdbCategory->create(
+        return $this->cmdbCategory->create(
             $workstationID,
             'C__CATG__LOGICAL_UNIT',
             [
@@ -195,15 +205,12 @@ abstract class BaseTest extends TestCase {
      * @throws \Exception
      */
     protected function addWorkstationComponent(int $workstationID, string $objectTypeConst): int {
-        $cmdbObject = new CMDBObject($this->api);
-        $cmdbCategory = new CMDBCategory($this->api);
-
-        $componentID = $cmdbObject->create(
+        $componentID = $this->cmdbObject->create(
             $objectTypeConst,
             $this->generateRandomString()
         );
 
-        return $cmdbCategory->create(
+        return $this->cmdbCategory->create(
             $componentID,
             'C__CATG__ASSIGNED_WORKSTATION',
             [
@@ -221,9 +228,7 @@ abstract class BaseTest extends TestCase {
      * @throws \Exception
      */
     protected function getIPv4Net(): int {
-        $cmdbObjects = new CMDBObjects($this->api);
-
-        return $cmdbObjects->getID('Global v4', 'C__OBJTYPE__LAYER3_NET');
+        return $this->cmdbObjects->getID('Global v4', 'C__OBJTYPE__LAYER3_NET');
     }
 
     /**
@@ -234,9 +239,7 @@ abstract class BaseTest extends TestCase {
      * @throws \Exception
      */
     protected function getRootLocation(): int {
-        $cmdbObjects = new CMDBObjects($this->api);
-
-        return $cmdbObjects->getID('Root location', 'C__OBJTYPE__LOCATION_GENERIC');
+        return $this->cmdbObjects->getID('Root location', 'C__OBJTYPE__LOCATION_GENERIC');
     }
 
     /**
@@ -249,9 +252,7 @@ abstract class BaseTest extends TestCase {
      * @throws \Exception
      */
     protected function addIPv4(int $objectID): int {
-        $cmdbCategory = new CMDBCategory($this->api);
-
-        return $cmdbCategory->create(
+        return $this->cmdbCategory->create(
             $objectID,
             'C__CATG__IP',
             [
@@ -276,9 +277,7 @@ abstract class BaseTest extends TestCase {
      * @throws \Exception
      */
     protected function defineModel(int $objectID): int {
-        $cmdbCategory = new CMDBCategory($this->api);
-
-        return $cmdbCategory->create(
+        return $this->cmdbCategory->create(
             $objectID,
             'C__CATG__MODEL',
             [
@@ -301,9 +300,7 @@ abstract class BaseTest extends TestCase {
      * @throws \Exception on error
      */
     protected function addObjectToLocation(int $objectID, int $locationID): int {
-        $cmdbCategory = new CMDBCategory($this->api);
-
-        return $cmdbCategory->create(
+        return $this->cmdbCategory->create(
             $objectID,
             'C__CATG__LOCATION',
             [
@@ -325,9 +322,7 @@ abstract class BaseTest extends TestCase {
      * @throws \Exception on error
      */
     protected function addContact(int $objectID, int $contactID, int $roleID = 1): int {
-        $cmdbCategory = new CMDBCategory($this->api);
-
-        return $cmdbCategory->create(
+        return $this->cmdbCategory->create(
             $objectID,
             'C__CATG__CONTACT',
             [
@@ -395,6 +390,16 @@ abstract class BaseTest extends TestCase {
         $formattedTimestamp = date('Y-m-d H:i:s', $timestamp);
         $this->assertInternalType('string', $formattedTimestamp);
         $this->assertSame($formattedTimestamp, $time);
+    }
+
+    /**
+     * Validate string as identifier
+     *
+     * @param int $value Positive integer
+     */
+    protected function isID(int $value) {
+        $this->assertInternalType('int', $value);
+        $this->assertGreaterThan(0, $value);
     }
 
     /**